+++ src/components/charts/index.ts
@@ -1,9 +1,0 @@
-"use client";
-
-export { TimeSeriesLineChart, BarChart, PieChart, Sparkline, MetricCard, FunnelChart } from "./quill";
-export type { TimeSeriesLineChartProps, TimeSeriesLineChartSeries } from "./quill";
-export type { BarChartProps, BarChartSeries } from "./quill";
-export type { PieChartProps, PieChartSlice } from "./quill";
-export type { SparklineProps } from "./quill";
-export type { MetricCardProps } from "./quill";
-export type { FunnelChartProps, FunnelStep } from "./quill/FunnelChart";